+++ client/src/component/common/Input.jsx
@@ -4,6 +4,7 @@ import { COLOR } from '../../style/theme';
 const Container = styled.div`
   display: flex;
   flex-direction: column;
+  width: 100%;
 `;
 
 const Label = styled.label`
